Added minimum free space

+++ clean.py
@@ -3,25 +3,35 @@ import os
 import shutil
 
 try:
-    from config import backups_dir, how_many_keep
+    from config import backups_dir, how_many_keep, min_free_gb
 except ImportError:
     raise ImportError('Create and adjust config.py: cp config.py.example config.py');
 
-dirs_list = []
+s = os.statvfs(backups_dir)
+free_space = (s.f_bavail * s.f_frsize) / 1024 / 1024 / 1024
 
-for dir, subdir, filenames in os.walk(backups_dir):
-    for dir in subdir:
-        if dir.isdigit():
-            dirs_list.append(dir)
+if min_free_gb > free_space:
+    print 'Found ' , free_space , ' free space on ' , backups_dir , ' witch is less than min_free_gb: ' , min_free_gb
 
-dirs_list.sort()
-# print dirs_list
+    dirs_list = []
 
-for i in range(how_many_keep):
-    if dirs_list:
-        dirs_list.pop()
+    for dir, subdir, filenames in os.walk(backups_dir):
+        for dir in subdir:
+            if dir.isdigit():
+                dirs_list.append(dir)
 
-print "Directories deleted: "
-print dirs_list
-for dir in dirs_list:
-    shutil.rmtree(backups_dir + dir)
+    dirs_list.sort()
+    # print dirs_list
+
+    for i in range(how_many_keep):
+        if dirs_list:
+            dirs_list.pop()
+
+    print 'Directories deleted: '
+    print dirs_list
+    for dir in dirs_list:
+        shutil.rmtree(backups_dir + dir)
+
+else:
+    print 'Found ',free_space,'GB free space on ' , backups_dir , ' witch is more than min_free_gb: ' , min_free_gb , 'GB.'
+    print 'No directories will be deleted.'
